wr65223 - radical changes to PRANG::Graph
Completely revamped the ugly API between PRANG::Marshaller and the
PRANG::Graph::* components; added PRANG::Graph::Context to handle it.  I
think this approach could be workable; the out-bound marshaller is the next
piece to write - and assuming the code doesn't start to smell too badly, it
should be ready for proving via tests.