wr65223 - move XML::EPP::EPP::* to XML::EPP::*, use PRANG::Graph API
The 'epp-1.0' API is really framework - the modules could go under
XML::EPP::Framework or something like that to keep the message classes out
of the way, but I've decided it looks nicer to avoid the double-up or
prefix, and that the downside - intermingling of message node classes and
sub-rfcs like Common, Domain, etc - is not such a big deal.

Move them about, brush them up a bit - remove lots of intermediate cruft
which was written as the PRANG::Graph API was taking shape.