wr62170 - fixes and more completion to PRANG
Filling out and fixing glaring bugs in the 'expected', 'complete' and
'accept' graph methods.  Instituting an exception class, for the dubious
practice of selecting behaviour based on the exception type.  It would be
better instead to add another method, which checks the next node but does
not move on like 'accept' does.