wr62170 - support PRANG::Graph-based roles in serializer: input/output
Adding support for attributes which may be roles.  In this implementation,
the objects must be PRANG::Graph classes; that is, they must have a defined
root element and be further validated.  When the graph is first calculated,
all of the classes which implement the role are treated as if they were
manually added to the node map.

Some rework of the (quite untested) xmlns support in this commit.