wr62170 - attributes with no namespace prefix are in *no* namespace
Whatever that means.  We'll interpret it as meaning that the namespace is
the same as the current element, treat it as such on input and suppress
them on output.  This will break when you want an attribute in a namespace
which has no prefix somewhere higher up in the document, but is in a
different namespace to the current element.  Later.