Delete SSL test script, add tests for and bugfix SRS::EPP::Proxy::Listener
Also delete the old ssl-fu test, tests for which now exist in
Net::SSLeay::OO itself.

It seems a shame that using 'IO::Socket::INET6' and passing a host name
does not listen on v4 addresses with that host name by default; then again,
it should return a single socket so the most specific one should win.  This
is something of an API design failure.  So, hostnames passed in get
converted to a list of addresses, all of which are listened on with
individual sockets.