rfc5734: use bytes for encoding safety
The 'length' of a string is the number of characters; not the number of
bytes.  Be sure to know which one is right.  There is a missing link, which
is making sure that the strings passed to the bytes:: functions are already
either A) binary-coded as returned by Encode::encode() or B) utf-8 valid.
This requires careful testing.