Fix-ups to Event logic
This isn't tested in the unit tests, unfortunately.  These changes were
found to be required by real testing etc.

Firstly, Event.pm wasn't even being loaded.  Secondly, the method of
starting the event loop was wrong.  Thirdly, output event watchers on
file handles need to be turned on and off, so that they are only active
when there is queued output data (due to a short write).  This is because
most of the time the handle is probably writable, we don't want an event
unless we have something to send.