Proxy: save the socket object returned by ->accept
As soon as that object falls out of scope, the IO::Socket::* destructor
goes ahead and closes the FD; as we passed the filehandle number directly
to the Net::SSLeay::OO::SSL object, it isn't being reference counted
against that Perl object.  So, save a copy of the raw socket to the Session
object.