Fix/monkeypatch daemonization
The reason we didn't seem to daemonize came down to the "close all open
FDs" part of daemonization.  MooseX::Daemonize::Core was doing this, and
this closed our accept socket.  Furthermore, the 'if $self->is_daemon' test
was missing, so basically the daemon process died and the starter process
started listening and servicing sessions!  Just goes to show, that
sometimes two fails make half a win... or something.