Input system: add debugging, and fix bugs
A few things - we can't rely on the input event to fire, as it is attached
to the input socket, which will be cleared by the SSL libraries as SSL
packets are read.  Instead, get the Packets::input_event to re-yield itself
if there is more input ready.

Also, the case where the input is empty (ie, the other end has hung up)
must be detected and a shutdown event passed back.