Session: add want_user
I'm not entirely sure this is the correct approach, but when dispatching
the message to the back-end, we still need to know which registrar ID to
set.  The effect of this change is to permit a queued login message to set
the effective registrar ID on any messages sent to the back-end; this is
safe working under the assumption that it is not possible to queue any
back-end transaction other than a login while not authenticated, and that
you cannot queue another login while authenticated.