Add some calls to unloop_all
I was working under the assumption that with no watchers ready, we'd drop
out of Event - but either there are some more watchers left I haven't found
or it will just wait for the timeout anyway.  So, explicitly unloop when
shutting down the session and if there is an exception.