Re-work the Session test script
The SRS::EPP::Session class had many changes that were made from in-situ
testing; many of these showed flaws in my assumptions, which I have had to
adjust the test script to allow for.  Along the way I made this
'loop_until' method, which could potentially be the core of a data-driven
test suite framework for mocked tests.