Login: cast input userid to an integer, return error if not
The login name must be at least 3 characters long according to the EPP
specification.  For SRS mapping, we are using the registrar ID as the
login name.  If the number is less than 100, then it therefore needs
to be padded with 0's - but the back-end does not like this.  So, cast
to a number (0+) as well as validating it is numeric.