XML Mapping Tests: sort out the multiple cycle test system
This makes sure that we're running the right assertions on the right stage,
as well as nicely labelling all of the cycles and unifying many of the
test names to be consistent.  Each notify/process cycle also gets a test
that it happened at all, which sections the tests nicely.  Also a design
bug is fixed in how new backend queue items are detected.