Error handling: deal with exceptions during PRANG validation
The code was already there, it just wasn't tested (nor did it really work
properly).  Fix it, using the tested functionality in XML::EPP.

Also update dependencies in Makefile.PL and debian/control (would be nice
if this could be more automatic) for the versions required to make this
work.