Revert "Session: unify mapping response code"
This reverts commit 0bf3566d6a0a03cb29d59e832ff985a3493a3d97.

The change doesn't seem to work... the problem is, it requires
proces_queue to be called multiple times for a single request,
which means the queue loses track of it's counter. It ends up
looking at a request that's not there, and the server hangs.

This will obviously need to be re-thought at a later date, but
reverting for now to allow testing to progress