wr65227: attempt to derive an error code if exceptions are passed in
If one or more XML::SRS::Error's are passed in to the constructor, and
no error code is passed, we map those errors to XML::EPP::Errors, and
derive the code. This is so we don't have to pass in some made up code
which will get replaced after the errors are mapped.