wr72782: return the upId element in the info domain response
It should only be returned if the domain has been updated at least once.
We derive this based on the registered date of the domain, and the audit
time of the response we get back. If they're the same, then it can't have
been updated. (The updated time also has this requirement).