wr72880: Fixed bug when doing check contact on multiple contacts
This seems to be the only command that does a query returning multiple 'results',
i.e. more than one handle from the handle details query. The handling of
this case called the 'response' method on XML::SRS::Result, which dies if
more than one response exists. We need to use the responses method instead.