wr72880 - rework the Response::Error class
Replace an imperative BUILDARGS with a set of lazy declarative functions.
This makes it tidy to customise the default result code for exceptions and
the like.  This means that the Session::input_packet function need not
specify 2001 to the result.  Add a couple of tests to demonstrate the
errors.  As we are asserting that the error message is more important than
the fine-grained error code, add tests for that too.