Session: remove a comment
I'm not sure about why this comment is here; the abstracted interface is
right above.  The job of the functions at that level are simply to write
whatever is ready to write, until an incomplete write happens.  As we are
in non-blocking mode, that means that the output is full.  So, we enable
the output event; this will make Event call output_event once the socket
is writable again, then when the buffer is empty, we turn off that event.