wr72027 - Session: use SRS unique identifiers for server_id (svTRID)
The intention was that this would be the case; the "sep.TIMEXXXX.PIDX.NNNN"
identifiers are temporary and only assigned if something asks for the
server_id before there is a back-end response (eg, if we have to generate
an ActionId).  So, assign them in the form X,NNNN as we are used to using.
In the case of batching the FE sequence ID will be made unique using [N],
where N is the part number of the response.  Be sure to log any changes.

This is to assist logging and diagnosis by making the identifiers as
consistent as possible across the different system components.