wr72027 - add stringification methods for SRS::EPP::(Command::Response)
First step at making log messages induce less conjunctival
haemorrhaging.  As all over the place, the objects are printed, add a
default stringification method.  Note this has the side effect of
assigning unique server IDs to *all* commands.

Also, make sure that the server_id default has a fallback if running
with no session (this happens in 04-backend-queue.t).

A good way to see the effect of this change to the logging is to run
mappings.t with -v; this prints to the terminal all of the INFO level
messages (use -vv for DEBUG as well).