wr73095 - rework logic around expiry dates
Also some structural simplification.  If you are using if() { } statements
that test preconditions, and the using "return" in the middle, there really
is little point in avoiding goto.  goto might be if's evil twin, but they
are still closely related.

This introduces a new error for the caught condition that the current
expiry date passed in is wrong.