[ws] workaround a limitation in perltidy.pl
MooseX::Method::Signatures already has the limitation that the
signature must be all on one line; the perltidy.pl script also assumes
that opening braces are on that line too.  These are the only two
declarations that don't, so rather than try to extend perltidy.pl
again, just change them.