Add a test that textNode exceptions are gracefully trapped
These were returning the internal MooseX::Method::Signatures exception
that was caused by the exception generation passing an unhandled LibXML
node to XML::EPP::Error::value; in fact the exception was coming from even
further down the stack.  With the temporally related changes in PRANG and
XML::EPP this is now handled gracefully.