wr73184 - restore do-nothing characteristic of default exception handler
This code block has a number of problems, primarily repeating the call
to write_to_client() with the same arguments each time through the
loop.  It also didn't include the length in the expected written
count, and was missing the utf8 conversion step found in
SRS::EPP::Session::send_reply.  These bugs are almost certainly why a
session which ends up in this state would reply repeatedly with the
same response, with eventually the SSL libraries throwing an exception
which kills the program.

However even with these bugs fixed it will not work; the protocol
state is completely ignored; it might be in the middle of issuing a
packet, or perhaps it already has an error response queued up for a
successful command, which has not been written yet, so you will reply
to the wrong command with this error.  It is simply not reasonable to
do anything except log the fault and exit.

The better solution is to make the individual event handlers more
robust.  That way, exceptions can be caught and error responses
gracefully placed in the reply queue.