[ws] Add parens around 'has' declarations
Normally I'm not a fan of throwing cosmetic fingernails all through
code to communicate a grouping which doesn't really exist (PBP makes
a similar case), but in this case without them, perltidy is doing
strange things with blocks inside the list.  So, we'd better use
them.  It has nothing to do with ridicule from the Moose community
about it, I assure you.