wr73426: return a 2500 error when the SRS response can't be parsed
This change required query id's being added to all SRS query transactions
(something that will probably assist in debugging anyway). This is so we
can easily find the command relating to an SRS request at any point in
the process. We can then return an error response to this command directly
rather than doing further processing.