wr72389 - revert to configurable timeouts
Also use the same event loop timeout as is specified in the config
file; it seems to be biting before the input timeout does.  The
underlying segfault issue which was attempted to be addressed by the
commits which hard-coded this value (referred to in commits 984e420c
and b47a9879) has hopefully been fixed.